% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_init}
\alias{eq_init}
\title{Initializes the database connection}
\usage{
eq_init(
  db_port = Sys.getenv("DB_PORT"),
  db_name = Sys.getenv("DB_NAME"),
  db_host = "localhost",
  db_user = Sys.getenv("USER"),
  log_level = logger::WARN
)
}
\arguments{
\item{db_port}{The port of the database to connect to. Defaults to the
environment variable "DB_PORT".}

\item{db_name}{The name of the database to connect to. Defaults to the
environment variable "DB_NAME".}

\item{db_host}{The host of the database to connect to. Defaults to the
"localhost". If the "DB_HOST" environment variable is set then that
will be used.}

\item{db_user}{The user name to connect to the database with. Defaults to
the current user name. If the "DB_USER" environment variable is set
then that will be used.}

\item{log_level}{the logging threshold level. Defaults to logger::WARN.}
}
\value{
TRUE on a successful initialization otherwise FALSE.
}
\description{
This must be called before using any of the functions in this package
}
