% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_pop_out_queue}
\alias{eq_pop_out_queue}
\title{Pops the highest priority task of the specified work type off
of the db out queue}
\usage{
eq_pop_out_queue(eq_type, delay = 0.5, timeout = 2)
}
\arguments{
\item{eq_type.}{Integer. The type of the work to pop from the queue}

\item{delay.}{Numeric. The initial polling delay value}

\item{timeout.}{Numeric. The duration after which this call will timeout
and return with an EQ_TIMEOUT result.}
}
\value{
A two element list whose first element is one of
ResultStatus$SUCCESS or ResultStatus$FAILURE. On success the
second element will be the popped eq task id. On failure, the second
element will be one of EQ_ABORT or EQ_TIMEOUT depending on the
cause of the failure.
}
\description{
This call repeatedly polls for a task of the specified type. The polling
interval is specified by
the delay such that the first interval is defined by the initial delay value
which is increased exponentionally after the first poll. The polling will
timeout after the amount of time specified by the timout value is has
elapsed.
}
