% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_push_in_queue}
\alias{eq_push_in_queue}
\title{Pushes the specified task onto the input queue.}
\usage{
eq_push_in_queue(eq_task_id, eq_type)
}
\arguments{
\item{eq_task_id}{Integer. The id of the task}

\item{eq_type}{Integer the type of the task}
}
\value{
ResultStatus$SUCCESS if the task was successfully pushed
onto the input queue, otherwise ResultStatus$FAILURE.
}
\description{
Pushes the specified task onto the input queue.
}
