% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_push_out_queue}
\alias{eq_push_out_queue}
\title{Pushes the specified task onto the output queue with
the specified priority.}
\usage{
eq_push_out_queue(eq_task_id, eq_type, priority = 0)
}
\arguments{
\item{eq_task_id}{Integer. The id of the task}

\item{eq_type}{Integer the type of the task}

\item{priority}{Integer the priority of the task}
}
\value{
ResultStatus$SUCCESS if the task was successfully pushed
onto the output queue, otherwise ResultStatus$FAILURE.
}
\description{
Pushes the specified task onto the output queue with
the specified priority.
}
