% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_query_result}
\alias{eq_query_result}
\title{Queries for the result of the specified task}
\usage{
eq_query_result(eq_task_id, delay, timeout)
}
\arguments{
\item{eq_task_id}{Integer. The id of the task to query}

\item{delay}{Numeric. The initial polling delay value}

\item{timeout}{Numeric.The duration after which the query will timeout}
}
\value{
A two element list whose first element indicates the status of
the query: ResultStatus$SUCCESS or ResultStatus$FAILURE, and whose
second element is either the result of the task, or in the case of
failure the reason for the failure ('EQ_TIMEOUT', or 'EQ_ABORT')
}
\description{
The query repeatedly polls for a result. The polling interval is specified by
the delay such that the first interval is defined by the initial delay value
which is increased exponentionally after the first poll. The polling will
timeout after the amount of time specified by the timout value is has
elapsed.
}
