% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_stop_worker_pool}
\alias{eq_stop_worker_pool}
\title{Stops any workers pools associated with the specified work type by
pusing EQ_STOP into the output queue.}
\usage{
eq_stop_worker_pool(eq_type)
}
\arguments{
\item{eq_type}{Integer. The work type for the pools to stop}
}
\description{
Stops any workers pools associated with the specified work type by
pusing EQ_STOP into the output queue.
}
