% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_submit_task}
\alias{eq_submit_task}
\title{Submits work of the specified type and priority with the specified
payload, returning the task id assigned to that task}
\usage{
eq_submit_task(exp_id, eq_type, payload, priority = 0)
}
\arguments{
\item{exp_id}{String. The id of the experiment of which the work is
part.}

\item{eq_type}{integer. The type of work}

\item{payload}{String. The work payload}

\item{priority}{integer. The priority of this work}
}
\value{
A list whose first element is the ResultStatus of the submission, and
whose second element is the task id assigned to this task if the submission
was successfull, otherwise EQ_ABORT.
}
\description{
Submits work of the specified type and priority with the specified
payload, returning the task id assigned to that task
}
