% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{queue_pop}
\alias{queue_pop}
\title{Performs the actual queue pop as defined the sql string
This call repeatedly attempts the pop operation by executing sql until
the operation completes or the timeout duration has passed. The polling
interval is specified by the delay such that the first interval is
defined by the initial delay value which is increased exponentionally
after the first poll. The polling will timeout after the amount of time
specified by the timout value is has elapsed.}
\usage{
queue_pop(sql_pop, delay, timeout)
}
\arguments{
\item{sql_pop}{String. The sql query that defines the pop operation}

\item{delay}{Numeric. The initial polling delay value}

\item{timeout}{Numeric. The duration after which this call will timeout
and return.}
}
\value{
A two element list where the first elements is one of
ResultStatus$SUCCESS or ResultStatus$FAILURE. On success the
second element will be the popped eq_task_id. On failure, the second
element will be one of 'EQ_ABORT' or 'EQ_TIMEOUT' depending on the
cause of the failure.
}
\description{
Performs the actual queue pop as defined the sql string
This call repeatedly attempts the pop operation by executing sql until
the operation completes or the timeout duration has passed. The polling
interval is specified by the delay such that the first interval is
defined by the initial delay value which is increased exponentionally
after the first poll. The polling will timeout after the amount of time
specified by the timout value is has elapsed.
}
