% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\docType{data}
\name{ResultStatus}
\alias{ResultStatus}
\title{Two element named list to indicate the status of the result task queries}
\format{
An object of class \code{list} of length 2.
}
\usage{
ResultStatus
}
\description{
ResultStatus$SUCCESS indicates a successful query and ResultStatus$FAILURE
indicates a failed query.
}
\keyword{datasets}
