% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{select_task_result}
\alias{select_task_result}
\title{Selects the result ('json_in') payload associated with the specified
task id in the eq_tasks table.}
\usage{
select_task_result(eq_task_id)
}
\arguments{
\item{eq_task_id}{Integer. The id of the task to get the json_in for.}
}
\value{
The result payload for the specified task id.
}
\description{
Selects the result ('json_in') payload associated with the specified
task id in the eq_tasks table.
}
