% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{SQL_update}
\alias{SQL_update}
\title{Utility function for updating the specified columns in the
specified table with the specified values.}
\usage{
SQL_update(table, col_names, values, where)
}
\arguments{
\item{table}{String. The name of the table to update}

\item{col_names}{List. A list of column names to update}

\item{values}{List. A list of values to update the columns with}

\item{where}{String. A SQL where clause, NOT including the "where"}
}
\value{
The number of rows effected by the update
}
\description{
Utility function for updating the specified columns in the
specified table with the specified values.
}
