% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{insert_task}
\alias{insert_task}
\title{Inserts the specified payload to the database, creating
a task entry for it and returning its assigned task id}
\usage{
insert_task(exp_id, eq_type, payload)
}
\arguments{
\item{exp_id}{Integer. The id of the experiment that this task is part of}

\item{eq_type}{Integer. The work type of this task}

\item{payload}{String. The task payload}
}
\value{
The task id assigned to this task.
}
\description{
Inserts the specified payload to the database, creating
a task entry for it and returning its assigned task id
}
